reader: extract useReaderChapter from ReaderPage
Pulled the 170-LOC fetch+normalize effect (cache lookup, public/userbook
fork, dedupe by fetchKey, wake-abort retry) into useReaderChapter.ts.
Page now calls one hook and gets {chapter, book, publicChapter, publicBook,
loading, error}.

ReaderPage: 976 → 778 LOC.
